WIP: interactivity api
